Create ImageDirectory to monitor and load photos from a directory
ImageDirectory starts a WatchService on the photo directory, and loads
all images for display usage.
-Note: Need to determine which method to use to recognize a photo from a
random file. Currently can either check the Mime Type, or try and load
the image to a BufferedImage. Mime Type is faster.
ImageDirectory also pauses all tasks in the Controller when any
specified pause file is found in the photo directory.